+++ src/interface/props.d.ts
@@ -1,0 +1,11 @@
+import { IProduct } from './main';
+
+export interface IProductCardProps {
+  product: IProduct;
+}
+
+export interface IProductModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  product: IProduct;
+}
